+++ src/index.ts
@@ -24,29 +24,6 @@ declare global {
   }
 }
 
-export function readJsonSchemas(): any[] {
-  const jsonDocuments: any[] = []
-  const pathToSchemaDir: string = '../schemas/v4'
-  const jsonFiles = fs
-    .readdirSync(pathToSchemaDir)
-    .filter((file) => path.extname(file) === '.json')
-  jsonFiles.forEach((file) => {
-    try {
-      // eslint-disable-next-line security/detect-non-literal-fs-filename
-      const fileData = fs.readFileSync(path.join(pathToSchemaDir, file), 'utf-8')
-      const jsonFile = JSON.parse(fileData.toString())
-      jsonDocuments.push(jsonFile)
-    } catch (err) {
-      OCEAN_NODE_LOGGER.log(
-        LOG_LEVELS_STR.LEVEL_WARN,
-        `Error loading DDO schema from ${path.join(pathToSchemaDir, file)}`,
-        true
-      )
-    }
-  })
-  return jsonDocuments
-}
-
 // we have 5 json examples
 // we should have some DDO class too
 function loadInitialDDOS(): any[] {
